requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.1
mccabe==0.6.1
Pillow==8.0.0
psycopg2==2.7.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.3.1
